requirements.txt
# Core ML stack
torch>=2.3.1
transformers>=4.44.2
datasets>=2.20.0
sentence-transformers>=3.0.1
scikit-learn>=1.3
numpy>=1.26
tqdm>=4.66

# Vector search
faiss-cpu>=1.8.0

# Document I/O
pymupdf>=1.24.10        # PDF parsing (imported as `fitz`)
python-docx>=1.1.2      # .docx parsing
pyyaml>=6.0             # YAML rule-packs

# API
fastapi>=0.110
uvicorn>=0.30

# Optional (only if you use adapters/LoRA in models/adapters/)
peft>=0.11.1
